# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Codex API",
    "description": "REST API for Codex digital library server",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/events/stream": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Subscribe to real-time entity change events via SSE",
        "description": "Clients can subscribe to this endpoint to receive real-time notifications\nabout entity changes (books, series, libraries) happening in the system.\n\n## Authentication\nRequires valid authentication with `LibrariesRead` permission.\n\n## Event Format\nEvents are sent as JSON-encoded `EntityChangeEvent` objects with the following structure:\n```json\n{\n  \"type\": \"book_created\",\n  \"book_id\": \"uuid\",\n  \"series_id\": \"uuid\",\n  \"library_id\": \"uuid\",\n  \"timestamp\": \"2024-01-06T12:00:00Z\",\n  \"user_id\": \"uuid\"\n}\n```\n\n## Keep-Alive\nA keep-alive message is sent every 15 seconds to prevent connection timeout.",
        "operationId": "entity_events_stream",
        "responses": {
          "200": {
            "description": "SSE stream of entity change events",
            "content": {
              "text/event-stream": {}
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "jwt_bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/filesystem/browse": {
      "get": {
        "tags": [
          "filesystem"
        ],
        "summary": "Browse filesystem directories",
        "description": "Returns a list of directories and files in the specified path",
        "operationId": "browse_filesystem",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "Path to browse (defaults to user's home directory)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Directory contents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/filesystem/drives": {
      "get": {
        "tags": [
          "filesystem"
        ],
        "summary": "Get system drives/volumes",
        "description": "Returns a list of available drives or mount points on the system",
        "operationId": "list_drives",
        "responses": {
          "200": {
            "description": "Available drives",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FileSystemEntry"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt_bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/api/v1/tasks/stream": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Subscribe to real-time task progress events via SSE",
        "description": "Clients can subscribe to this endpoint to receive real-time notifications\nabout background task progress (analyze_book, generate_thumbnails, etc.).\n\n## Authentication\nRequires valid authentication with `LibrariesRead` permission.\n\n## Event Format\nEvents are sent as JSON-encoded `TaskProgressEvent` objects with the following structure:\n```json\n{\n  \"task_id\": \"uuid\",\n  \"task_type\": \"analyze_book\",\n  \"status\": \"running\",\n  \"progress\": {\n    \"current\": 5,\n    \"total\": 10,\n    \"message\": \"Processing book 5 of 10\"\n  },\n  \"started_at\": \"2024-01-06T12:00:00Z\",\n  \"library_id\": \"uuid\"\n}\n```\n\n## Keep-Alive\nA keep-alive message is sent every 15 seconds to prevent connection timeout.",
        "operationId": "task_progress_stream",
        "responses": {
          "200": {
            "description": "SSE stream of task progress events",
            "content": {
              "text/event-stream": {}
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "jwt_bearer": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrowseResponse": {
        "type": "object",
        "required": [
          "current_path",
          "entries"
        ],
        "properties": {
          "current_path": {
            "type": "string",
            "description": "Current directory path"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileSystemEntry"
            },
            "description": "List of entries in the current directory"
          },
          "parent_path": {
            "type": [
              "string",
              "null"
            ],
            "description": "Parent directory path (None if at root)"
          }
        },
        "example": {
          "current_path": "/home/user/Documents",
          "entries": [
            {
              "is_directory": true,
              "is_readable": true,
              "name": "Comics",
              "path": "/home/user/Documents/Comics"
            },
            {
              "is_directory": true,
              "is_readable": true,
              "name": "Manga",
              "path": "/home/user/Documents/Manga"
            }
          ],
          "parent_path": "/home/user"
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "details": {},
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "FileSystemEntry": {
        "type": "object",
        "required": [
          "name",
          "path",
          "is_directory",
          "is_readable"
        ],
        "properties": {
          "is_directory": {
            "type": "boolean",
            "description": "Whether this is a directory"
          },
          "is_readable": {
            "type": "boolean",
            "description": "Whether the entry is readable"
          },
          "name": {
            "type": "string",
            "description": "Name of the file or directory"
          },
          "path": {
            "type": "string",
            "description": "Full path to the entry"
          }
        },
        "example": {
          "is_directory": true,
          "is_readable": true,
          "name": "Documents",
          "path": "/home/user/Documents"
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "jwt_bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login"
      }
    }
  },
...
}